reorganize ideas into proper folders update readme
