+++ src/ideas/environment/Background.tsx
