+++ src/ideas/environment/Fog.tsx
