+++ src/ideas/index.ts
@@ -6,14 +6,14 @@ export * from "./modifiers/Floating";
 export * from "./modifiers/LookAtPlayer";
 export * from "./modifiers/Spinning";
 export * from "./modifiers/Tool";
-export * from "./physical/Image";
-export * from "./physical/Arrow";
-export * from "./physical/Audio";
-export * from "./physical/Background";
-export * from "./physical/Fog";
-export * from "./physical/Frame";
-export * from "./physical/HDRI";
-export * from "./physical/InfinitePlane";
-export * from "./physical/Model";
-export * from "./physical/Video";
-export * from "./ux/TextInput";
+export * from "./media/Image";
+export * from "./ui/Arrow";
+export * from "./media/Audio";
+export * from "./environment/Background";
+export * from "./environment/Fog";
+export * from "./mediated/Frame";
+export * from "./media/HDRI";
+export * from "./environment/InfinitePlane";
+export * from "./media/Model";
+export * from "./media/Video";
+export * from "./ui/TextInput";
