+++ src/ideas/media/Model.tsx
