+++ src/ideas/ui/TextInput/logic/blink.ts
